cleanup

+++ elicitor.py
@@ -160,7 +160,7 @@ if family == 'metalog':
 
 	r_x = robjects.FloatVector([q for p, q in quantiles])
 	r_probs = robjects.FloatVector([p for p, q in quantiles])
-	r_term_limit = robjects.IntVector([term])
+	r_term_limit = robjects.FloatVector([term])
 	r_n_samples = robjects.IntVector([nsamples])
 	r_step_len = robjects.FloatVector([step_len])
 	r_bounds = robjects.FloatVector(bounds)
